// shared/utils/games.generated.ts
export interface GameVariant {
  id: number
  key: string
  group: number
  name: string
  shortName: string
  variantName: string
  hidden: boolean
  custom: boolean
  usesSeed: boolean
}

export interface GameGroup {
  id: number
  name: string
  variants: number[]
}